requirements.txt
appdirs==1.4.0
blinker==1.4
crcmod==1.7
Django==1.10.5
django-ipware==1.1.6
django-rest-framework-mongoengine==3.3.1
django-simple-captcha==0.5.3
djangorestframework==3.5.3
djangorestframework-gis==0.11
djangorestframework-jwt==1.9.0
mongoengine==0.11.0
olefile==0.44
oss2==2.2.3
packaging==16.8
Pillow==4.0.0
py==1.4.32
pydenticon==0.3
PyJWT==1.4.2
pymongo==3.4.0
pyparsing==2.1.10
pytest==3.0.6
requests==2.12.5
six==1.10.0
